mysensors.js: the presentation message does not hold the library version
Though https://www.mysensors.org/download/serial_api_20 states that
'The payload of presentation message will be set to the library version
(node device) or an optional description for the sensors.', no
indicator is available to let the receiver choose between these two
values.
More https://www.mysensors.org/download/sensor_api_20 states that the
payload should be 'an optional textual description of the attached sensor'.
So rather relies on the C_INTERNAL command to get the library version.